Add files via upload
These matlab scripts have the following functions:

Relaxation_matrix... creates a spin relaxation matrix by numerical integration and compares it in a plot to a Redfield matrix calculated on a single selected frame.  It parses an AMBER MD trajectory to extract a subset of coordinates using some routines from the mdtoolbox of Y matsunaga.  It reads additional parameters it needs from an Excel workbook and writes some derived parameters to this workbook.  Workbook access is provided by the Prepare_xlsx script. Other scripts use the workbook for input and output.  An example workbook is attached.  It then calls the ngce_IK_JHP3 script to do numerical integrations on coordinates from the trajectory segment by segment.

The two noe_spin_cluster scripts are modeled on noe_sprits from the Spinach library.  JHP_7 observes a selected spin and inverts selectively all other spins in the cluster.  JHP_8 inverts a selected spin and observes all other spins in the cluster.  Both write NOEs to the prepared Excel workbook for a list of provided mixing times.