Update README.md
The attached scripts have the following functions:

Relaxation_matrix...  calculates a relaxation matrix in a spherical tensor basis set using coordinates from an MD trajectory and other parameters read from an Excel workbook.  It uses routines from Matsunaga's mdtoolbox to select a subset of coordinates from the trajectory.  It calls the Prepare_xlsx.. script to read and write derived parameters to the workbook.  It then calls the ngce... script to calculate elements of the matrix using numerical integration routines segment by segment on the trajectory.  It calculates a Redfield matrix based on a single selected frame and plots a comparison of Redfield and numerical matrices.  Both matrices are saved for future use. A example workbook is included.

The noe_... scripts use the matrices and data from the workbook to calculate selective inversion NOEs in a manner modeled on other NOE scripts in the Spinach package.  One script, ..7 observes a selected spin from the subset of coordinates and successively inverts all other spins.  the  ..8 script inverts the selected spin and observes all others.  The calculated NOEs for a list of mixing times are written to the workbook.